Extract `Stream`-specific code into a separate file
